fix: dtype in action initialization

+++ envs/adapters/brax.py
@@ -27,7 +27,7 @@ class Brax(Environment):
     def reset(self, key: PRNGKeyArray) -> Tuple[Transition, EnvInfo, BraxEnvState]:
         env_state = self.env.reset(key)
         transition = Transition(
-            action=jnp.zeros(self.action_space.shape),
+            action=jnp.zeros(self.action_space.shape, dtype=self.action_space.dtype),
             next_obs=env_state.obs,
             reward=jnp.zeros(()),
             done=jnp.zeros((), dtype=bool),
